+++ src/admin/components/RadioInputField.vue
@@ -9,6 +9,11 @@
           <label class="block" for="default">Default</label>
         </div>
         <div class="text-center">
+          <input class="block" type="radio" id="select" value="select"
+          v-model="field.data.type" @change="onChange">
+          <label class="block" for="select">Select</label>
+        </div>
+        <div class="text-center">
           <input class="block" type="radio" id="color" value="color"
           v-model="field.data.type" @change="onChange">
           <label class="block" for="color">Color</label>