+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+### v2.2.0
+
+1. Features
+	- Personalize Forms now can be loaded via "product_tag"
+	- TextInputField now has max length, set to -1 so use no limit
+	- Add Select type for Radio input
+1. Fixes
+	- Fix meta 'name' index bug
+
 ### v2.1.0
 
 1. Features